[fix] GameEngine: rewrite user interaction
What?: Move all the classes and modules outside main

Breaking Changes? there is no game logic
